+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [NEXT_RELEASE]
+### Added
+- read file `.teresaignore` to ignore some files in deploy process
+
 ### Changed
 - some code refactors
 - default client timeout to 30 minutes